Tidied up and fixed some inconsistencies

+++ scripts/start-workflow.sh
@@ -4,13 +4,15 @@ set -Ee
 # Uncomment to enable debugging
 #set -vo xtrace
 
-#source $(dirname $0)/functions.sh
-#source $(dirname $0)/config.sh
 source scripts/functions.sh
 source settings/settings.conf
 
 # Run the die function to clean up files and give some troubleshooting help in case the workflow is killed or fails for some reason
 trap die ERR SIGTERM SIGINT
+
+# Set default variables
+START=$(date +%s)
+DATE=$(date +%F.%H.%M.%S)
 
 # Print the Selma ascii serpent
 selma-ascii
@@ -19,9 +21,11 @@ selma-ascii
 die () {
 	printf "#########################################\n"
 	err "$0 failed at line $BASH_LINENO"
-	warn "Will now attempt to delete $STAGINGDIR/$WFDIRNAME-$DATE"
-	rm -r $STAGINGDIR/$WFDIRNAME-$DATE && \
-	warn "$STAGINGDIR/$WFDIRNAME-$DATE has been deleted." || err "Deleting $INTERMEDSTOR/$WFDIRNAME-$DATE failed. Try deleting it manually."
+	if [[ -d $STAGINGDIR/$WFDIRNAME-$DATE ]]; then
+		warn "Will now attempt to delete $STAGINGDIR/$WFDIRNAME-$DATE"
+		rm -r $STAGINGDIR/$WFDIRNAME-$DATE && \
+		warn "$STAGINGDIR/$WFDIRNAME-$DATE has been deleted." || err "Deleting $STAGINGDIR/$WFDIRNAME-$DATE failed. Try deleting it manually."
+	fi
 
 	if [[ $SLURMID == [0-9]* ]]; then
 		warn "Stopping slurm job $SLURMID"
@@ -35,7 +39,15 @@ die () {
 		inf "Here is the slurm job file: ${OUTPUTDIR}/$WFDIRNAME-$DATE/slurm-$SLURMID.out"
 	fi
 
-	err "Something seems to have gone wrong, here's the logfile: $LOGFILE"
+	# Print benchmarking data
+	FINISH=$(date +%s)
+	EXECTIME=$(( $FINISH-$START ))
+	succ "Finished on $(date)"
+	printf "[$(showdate)][INFO]: Workflow duration: %dd:%dh:%dm:%ds\n" $((EXECTIME/86400)) $((EXECTIME%86400/3600)) $((EXECTIME%3600/60)) $((EXECTIME%60))
+
+	if [[ -f $LOGFILE ]]; then
+		err "Something seems to have gone wrong, here's the logfile: $LOGFILE"
+	fi
 	printf "#########################################\n"
 	exit 1
 }
@@ -44,10 +56,6 @@ die () {
 if [[ ! $@ ]]; then
 	usage
 fi
-
-# Set default variables
-START=$(date +%s)
-DATE=$(date +%F.%H.%M.%S)
 
 while getopts 'i:t:o:r:l:h' flag; do
 	case "${flag}" in
@@ -138,8 +146,8 @@ if [[ -z "${r}" ]]; then
 fi
 
 # make a log file at $OUTPUTDIR/$WFDIRNAME-$DATE/ named logfile.log
-mkdir $OUTPUTDIR/$WFDIRNAME-$DATE
-LOGFILE=$OUTPUTDIR/$WFDIRNAME-$DATE/logfile.log
+mkdir -p $OUTPUTDIR/logfiles/
+LOGFILE=$OUTPUTDIR/logfiles/logfile-$DATE.log
 exec &> >(tee $LOGFILE)
 
 # Check that the supplied input file folder contains supported input files before proceeding
@@ -177,22 +185,22 @@ if [[ -n "${INTERVAL:-}" ]]; then
 	fi
 else
 	if [[ "${HGVER}" == hg38 ]]; then
-		INTERVAL="${REFERENCES}"/"${HGVER}"/wgs_calling_regions.hg38.interval_list
+		INTERVAL="${REFERENCES%/}"/"${HGVER}"/wgs_calling_regions.hg38.interval_list
 	elif [[ "${HGVER}" == b37 ]]; then
-		INTERVAL="${REFERENCES}"/"${HGVER}"/b37_wgs_calling_regions.v1.interval_list
+		INTERVAL="${REFERENCES%/}"/"${HGVER}"/b37_wgs_calling_regions.v1.interval_list
 	fi
 fi
 
 # Make fresh copy of the workflow directory in case multiple executions of the workflow are started simultaneously
-inf "Copying $WFDIR to staging directory: $STAGINGDIR/$WFDIRNAME-$DATE\n"
+inf "Copying $WFDIR to staging directory: $STAGINGDIR/$WFDIRNAME-$DATE"
 rprog $WFDIR/ $STAGINGDIR/$WFDIRNAME-$DATE
 
 # Copy TSV file to $STAGINGDIR/$WFDIR-$DATE/workspace/samples.tsv
-inf "Copying $TSV to $STAGINGDIR/$WFDIRNAME-$DATE/workspace/samples.tsv\n"
+inf "Copying $TSV to $STAGINGDIR/$WFDIRNAME-$DATE/workspace/samples.tsv"
 rprog $TSV $STAGINGDIR/$WFDIRNAME-$DATE/workspace/samples.tsv
 
 # Say that the input files/folder is being copied to $STAGINGDIR/$WFDIRNAME-$DATE/workspace/ on the /cluster disk
-inf "Copying input files to $STAGINGDIR/$WFDIRNAME-$DATE/workspace/ this may take a while if your input files are large.\n"
+inf "Copying input files to $STAGINGDIR/$WFDIRNAME-$DATE/workspace/ this may take a while if your input files are large."
 rprog $INPUTS/* "$STAGINGDIR/$WFDIRNAME-$DATE"/workspace/
 
 # Run sbatch with all variables set and store the slurm id in $SLURMID
@@ -210,7 +218,7 @@ if [[ $SLURMID == '' ]]; then
 	die
 fi
 
-succ "Workflow with slurm ID $SLURMID has been submitted to the Colossus slurm queue\n"
+succ "Workflow with slurm ID $SLURMID has been submitted to the Colossus slurm queue"
 
 # Set to 0 so that the $STATUS variable is checked the first time the $STATUS variable is created
 PREVSTAT=0
@@ -256,7 +264,7 @@ while true; do
 	elif [[ -e $JOBSUCCESS ]]; then
 		succ "File transfer trigger from slurm job with ID $SLURMID has been detected, the file transfer has completed."
 		sleep 1
-		inf "Will now attempt to copy output files from $STAGINGDIR/$WFDIRNAME-$DATE/workspace/Outputs/ to ${OUTPUTDIR}/$WFDIRNAME-$DATE\n"
+		inf "Will now attempt to copy output files from $STAGINGDIR/$WFDIRNAME-$DATE/workspace/Outputs/ to ${OUTPUTDIR}/$WFDIRNAME-$DATE"
 		sleep 5
 		break
 	fi
@@ -310,30 +318,28 @@ while true; do
 done
 
 # Only copy the final bam file that goes into HaplotypeCaller as well as the VCF files from ApplyVqsr
-############ And temporarily also the benchmark files
-rprog $STAGINGDIR/$WFDIRNAME-$DATE/workspace/Outputs/benchmarks $INTERMEDSTOR/$WFDIRNAME-$DATE/workspace/Outputs/GatherBamFiles $INTERMEDSTOR/$WFDIRNAME-$DATE/workspace/Outputs/ApplyVqsr* ${OUTPUTDIR}/$WFDIRNAME-$DATE/ && \
+rprog $STAGINGDIR/$WFDIRNAME-$DATE/workspace/Outputs/GatherBamFiles $STAGINGDIR/$WFDIRNAME-$DATE/workspace/Outputs/ApplyVqsr* ${OUTPUTDIR}/$WFDIRNAME-$DATE/ && \
 succ "Files have been copied from $STAGINGDIR/$WFDIRNAME-$DATE/workspace/Outputs/ to ${OUTPUTDIR}/$WFDIRNAME-$DATE"
 
 # Copy the slurm file to the final output directory
-#################### Will this work if the wf is started from any directory? Does the code need to be made specific to something like $(pwd) or something?
 inf "Copying slurm-$SLURMID.out to ${OUTPUTDIR}/$WFDIRNAME-$DATE"
 rprog $WFDIR/slurm-$SLURMID.out ${OUTPUTDIR}/$WFDIRNAME-$DATE/slurm-$SLURMID.out
-#####################
 
 # Clean up any intermediary files
 inf "Attempting to delete staging directory: $STAGINGDIR/$WFDIRNAME-$DATE"
 rm -r $STAGINGDIR/$WFDIRNAME-$DATE && \
-succ "Staging directory $STAGINGDIR/$WFDIRNAME-$DATE has been deleted\n"
+succ "Staging directory $STAGINGDIR/$WFDIRNAME-$DATE has been deleted"
 
 # All done!
 if [[ ! -z "$(ls "${OUTPUTDIR}/$WFDIRNAME-$DATE")" ]]; then
-        succ "Your output files from slurm job $SLURMID are in ${OUTPUTDIR}/$WFDIRNAME-$DATE"
+	succ "Your output files from slurm job $SLURMID are in ${OUTPUTDIR}/$WFDIRNAME-$DATE"
 	succ "Workflow execution is all done!"
 else
-        err "The output directory "${OUTPUTDIR}/$WFDIRNAME-$DATE" from slurm job $SLURMID seems to be empty, something is broken, sorry. Please contact oskar.vidarsson@uib.no for support. \n"
+	err "The output directory "${OUTPUTDIR}/$WFDIRNAME-$DATE" from slurm job $SLURMID seems to be empty, something is broken, sorry. Please contact oskar.vidarsson@uib.no for support."
 fi
 
 # Print benchmarking data
 FINISH=$(date +%s)
 EXECTIME=$(( $FINISH-$START ))
-printf "The entire workflow, including input file transfer, job execution on Colossus, and output file transfer back to the p172ncspmdata disk, took %dd:%dh:%dm:%ds\n" $((EXECTIME/86400)) $((EXECTIME%86400/3600)) $((EXECTIME%3600/60)) $((EXECTIME%60))
+succ "Finished on $(date)"
+printf "[$(showdate)][INFO]: Total workflow duration: %dd:%dh:%dm:%ds\n" $((EXECTIME/86400)) $((EXECTIME%86400/3600)) $((EXECTIME%3600/60)) $((EXECTIME%60))
